allow at least one known operand in url
URL FORMAT: <host>/problem/subtraction/hard/50
there is something strange going on with the parameter collection and
the type coercion. Operands are expected as an array, but params returns
strings. Need to go through and change all operands to numbers and fit
them in an array. This is probably happening at the wrong place in the
application